Add script to build i386 libomp.dylib
dylib still links against /usr/lib/libSystem.dylib, so not sure if this
is totally correct.

.gitignore was added to ignore the generated build/ directory containing
the CMake out-of-source openmp build.

.gitmodules was edited because the openmp build copies its results to
the openmp repo runtime/exports directory, which makes the submodule
dirty to git. The submodule shouldn't have its sources changed, so git
shouldn't need to worry about anything that happens in there.